feat: Adds express server and functions for interacting with apis
This commit creates the Express.js server in src/server/index.js.
It adds the fully-functional searchGeoNamesByLocation function
which is able to communicate with the GeoNames API. Additionally,
it adds the dotenv package. Environment variables are stored in
the root level `.env` file. This also adds the express-server
script to package.json

Related to: #9