feat: Implements fully functional backend express server
This commit finishes #9 by implementing the backend server in
express.js. The server consists of three main parts:
- fetching geo data in file geoNamesSearch from the geonames.org api.
- fetching current or predicted weather for the location, with info
  from latitude and longitude data retrieved from the geonames api.
- fetching a picture of the location from pixabay.

All authentication is located in the root-level `.env` file.
Additionally, handleFormSubmit on the client side has been
expanded to successfully send date and location information
to the backend server and receive the response.

Resolves: #9